% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{change_layout}
\alias{change_layout}
\title{change_layout}
\usage{
change_layout(phytree, layout)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}

\item{layout}{(Required.)  One of 'radial', 'rect', 'unrooted', 'clock'.}
}
\description{
change the layout
}
