% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{change_tipsize}
\alias{change_tipsize}
\title{change_tipsize}
\usage{
change_tipsize(phytree, size = 10)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}
}
\description{
change tipsize
}
