% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_colormaps.R
\name{create_colormaps}
\alias{create_colormaps}
\title{create_colormaps}
\usage{
create_colormaps(df, custommaps = NULL, exclude_attr = c("node", "ancestor",
  "node.type"))
}
\arguments{
\item{df}{Required. A dataframe containing information about nodes in the tree.}

\item{exclude_attr}{a list of columns to exclude from the final output. The
defaults are the otputs from phylobase}

\item{custommaps.}{Optional. Curently a stub. There needs to be a way to pass
in custom colors.}
}
\description{
create colormaps for dataframe character/factor columns
}
