% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{hide_controls}
\alias{hide_controls}
\title{hide_controls}
\usage{
hide_controls(phytree)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}
}
\description{
hide the controls
}
