% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{highlight_tips_list}
\alias{highlight_tips_list}
\title{highlight_tips_list}
\usage{
highlight_tips_list(phytree, nodenames, column = "strain",
  highlight_color = NULL, highlight_color_stroke = NULL,
  highlight_size = NULL, highlight_stroke_width = NULL, add_dash = FALSE,
  opacity = NULL)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}

\item{nodenames}{(Required.) a character vector of tips.}

\item{column}{(Required.) the column/attribute used to search for those names.}

\item{highlight_color}{(Optional.) a six character hex color, (e.g. "#60AA9E")}

\item{highlight_color_stroke}{(Optional.) a six character hex color, (e.g. "#60AA9E")}

\item{highlight_size}{(Optional.) integer for sizing tips}

\item{highlight_stroke_width}{(Optional.) number for sizing branch widths}

\item{add_dash}{(Optional.) add a dash to the lines}

\item{opacity}{(Optional.) provide opacity (0-1).}
}
\description{
this function allows you to pass options for tips to be highlighted using a membership list
}
