% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree.R
\name{phylotree}
\alias{phylotree}
\title{phylotree html widget}
\usage{
phylotree(tree, data = NULL, layout = "radial", distance = "div",
  tipFontSize = 8, orientation_x = 1, orientation_y = 1, margins = 70,
  zoomLevel_x = 1, zoomLevel_y = 1, pan_x = 0, pan_y = 0,
  tipRadius = 10, tipMinRadius = 1, tipMaxRadius = 15,
  tipStroke = "#555", tipFill = "#555", tipStrokeWidth = 1,
  branchStroke = "#555", branchStrokeWidth = 2, branchOpacity = 0.3,
  autoTipSize = TRUE, width = NULL, height = NULL, elementId = NULL,
  highlight_color = "#FF0000", highlight_size = 30, python = "python",
  use_python = FALSE)
}
\arguments{
\item{tree}{(Required.)  a tree of type \code{\link[ape]{phylo}} or
\code{\link[phylobase]{phylo4}} or  \code{\link[phylobase]{phylo4d}}}

\item{data}{(Required.) Either a data.frame or NULL. This contains metadata
            about the tips of the tree


\seealso{
\href{'https://github.com/nextstrain/phyloTree'}
}}
}
\description{
create an interactive phylogenetic tree.
}
