% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree.R
\name{phylotree_html}
\alias{phylotree_html}
\title{This is the HTML Widget custom render html fn.}
\usage{
phylotree_html(id, style, class, width, height, ...)
}
\description{
This is the HTML Widget custom render html fn.
}
