% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{scale_branchthickness}
\alias{scale_branchthickness}
\title{scale_branchthickness}
\usage{
scale_branchthickness(phytree, scale = "linear", size = NULL)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}
}
\description{
change scale_branchthickness
}
