% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifications.R
\name{show_controls}
\alias{show_controls}
\title{show_controls}
\usage{
show_controls(phytree)
}
\arguments{
\item{phytree}{(Required.)  a phylotree \code{\link[htmlwidgets]{htmlwidget}}}
}
\description{
show the controls
}
