% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotree.R
\name{tiplabel_control_switch}
\alias{tiplabel_control_switch}
\title{Using Bulma Extension CSS for Switch Behavior}
\usage{
tiplabel_control_switch(id)
}
\description{
Using Bulma Extension CSS for Switch Behavior
}
